Added multiple Postman Runner tests

+++ www/app/Controllers/TasksController.php
@@ -1,35 +1,54 @@
 <?php namespace App\Controllers;
 
 use CodeIgniter\Events\Events;
-use App\Models\DocumentModel;
 use App\Models\TaskModel;
 use App\Models\StackModel;
-use App\Models\TaskAssigneeModel;
 use App\Models\TaskWatcherModel;
-use App\Models\TaskExtensionModel;
+use App\Models\UserModel;
 
 class TasksController extends BaseController
 {
     protected $permissionSection = "tasks";
 
+    private function getTask($taskId)
+    {
+        helper("permissions");
+
+        $taskModel = new TaskModel();
+        $task = $taskModel
+            ->select("tasks.*")
+            ->join("permissions", "permissions.resource = tasks.id AND permissions.user = ".$this->db->escape($this->request->user->id), 'left')
+            ->groupStart()
+                ->where("public", 1)
+                ->orGroupStart()
+                    ->where("public", 0)
+                    ->where("owner", $this->request->user->id)
+                ->groupEnd()
+                ->orWhere("permissions.permission IS NOT NULL", null)
+            ->groupEnd()
+            ->find($taskId);
+
+        permissions_load_permission($task, $this->request->user->id);
+
+        return $task;
+    }
+
     public function one_v1($taskId)
     {
-        // helper("tasks");
-        // $task = task_load($taskID);
-        $taskModel = new TaskModel();
-        $task = $taskModel->find($taskId);
+        $task = $this->getTask($taskId);
 
         if (!$task) {
             return $this->reply(null, 404, "ERR-TASKS-NOT-FOUND");
         }
 
-        $documentModel = new DocumentModel();
-        $documentModel->user = $this->request->user;
-        $document = $documentModel->find($task->project);
+        $document = $this->getDocument($task->project);
 
         if (!$document) {
             return $this->reply(null, 404, "ERR-TASKS-NOT-FOUND");
         }
+
+        $this->can("read", $document);
+        $this->can("read", $task);
 
         $stackModel = new StackModel();
         $stack = $stackModel->find($task->stack);
@@ -38,13 +57,41 @@ class TasksController extends BaseController
             return $this->reply(null, 404, "ERR-TASKS-NOT-FOUND");
         }
 
-        // return $this->reply(task_format($task));
         return $this->reply($task);
     }
 
     public function add_v1($stackId)
-    {
-        $taskData = $this->request->getJSON();
+    {        
+        $stackModal = new StackModel();
+        $stack = $stackModal->find($stackId);
+
+        if (!$stack) {
+            return $this->reply("Stack not found", 404, "ERR-TASK-CREATE");
+        }
+
+        $document = $this->getDocument($stack->project);
+
+        if (!$document) {
+            return $this->reply("Project not found", 404, "ERR-TASK-CREATE");
+        }
+
+        $this->can("add", $document);
+
+        $data = $this->request->getJSON();
+        $data->updated = null;
+        $data->archived = null;
+        $data->completed = null;
+        $data->project = $document->id;
+        $data->stack = $stack->id;
+        $data->position = 1;
+        // by default the owner is the user creating the task
+        $data->owner = $this->request->user->id;
+        $data->public = 1;
+
+        $taskModel = new TaskModel();
+        $taskModel->formatData($data);
+
+        // getting the desired position
         $position = $this->request->getGet("position");
 
         if (!$position) {
@@ -53,46 +100,6 @@ class TasksController extends BaseController
 
         if (!in_array($position, ["top", "bottom"])) {
             return $this->reply("Invalid task position", 500, "ERR-TASK-CREATE");
-        }
-        
-        $stackModal = new StackModel();
-        $stack = $stackModal->find($stackId);
-
-        if (!$stack) {
-            return $this->reply("Stack not found", 404, "ERR-TASK-CREATE");
-        }
-
-        $user = $this->request->user;
-        $documentModel = new DocumentModel();
-        $documentModel->user = $this->request->user;
-        $document = $documentModel->find($stack->project);
-
-        if (!$document) {
-            return $this->reply("Project not found", 404, "ERR-TASK-CREATE");
-        }
-
-        $this->can("add", $document);
-
-        // enforce an id in case there"s none
-        if (!isset($taskData->id)) {
-            helper("uuid");
-            $taskData->id = uuid();
-        }
-
-        $taskModel = new TaskModel();
-
-        $taskData->updated = null;
-        $taskData->archived = null;
-        $taskData->completed = null;
-        $taskData->project = $document->id;
-        $taskData->stack = $stack->id;
-        $taskData->position = 1;
-        // by default the owner is the user creating the task
-        $taskData->owner = $user->id;
-        $taskData->public = 1;
-
-        if (isset($taskData->repeats)) {
-            $taskData->repeats = \json_encode($taskData->repeats);
         }
 
         // get the last order number in from that project and stack
@@ -103,7 +110,7 @@ class TasksController extends BaseController
                 ->first();
             
             if ($lastPosition) {
-                $taskData->position = intval($lastPosition->position) + 1;
+                $data->position = intval($lastPosition->position) + 1;
             }
 
         // move all tasks order up by 1
@@ -117,7 +124,7 @@ class TasksController extends BaseController
         }
         
         try {
-            if ($taskModel->insert($taskData) === false) {
+            if ($taskModel->insert($data) === false) {
                 $errors = $taskModel->errors();
                 return $this->reply($errors, 500, "ERR-TASK-CREATE");
             }
@@ -125,10 +132,14 @@ class TasksController extends BaseController
             return $this->reply($e->getMessage(), 500, "ERR-TASK-CREATE");
         }
 
-        $task = $taskModel->find($taskData->id);
+        // inserting the default document permission
+        try {
+            $this->addPermission($data->id, $this::PERMISSION_TYPE_TASK);
+        } catch (\Exception $e) {
+            return $this->reply($e->getMessage(), 500, "ERR-TASK-CREATE");
+        }
 
-        helper("tasks");
-        $task = task_format($task);
+        $task = $taskModel->find($data->id);
 
         $this->addActivity(
             $document->id,
@@ -141,24 +152,35 @@ class TasksController extends BaseController
         return $this->reply($task);
     }
 
-    public function update_v1($taskID)
+    public function update_v1($taskId)
     {
-        $this->lock($taskID);
+        $this->lock($taskId);
+        $task = $this->getTask($taskId);
+
+        $data = $this->request->getJSON();
         
-        $user = $this->request->user;
         $taskModel = new TaskModel();
-        $task = $taskModel->find($taskID);
+        $taskModel->formatData($data);
+        unset($data->id);
+        unset($data->position);
+        unset($data->project);
+        unset($data->stack);
+        unset($data->assignees);
+        unset($data->info);
+        $data->archived = null;
 
         if (!$task) {
             return $this->reply(null, 404, "ERR-TASKS-NOT-FOUND");
         }
 
-        helper("documents");
-        $document = documents_load_document($task->project, $user);
+        $document = $this->getDocument($task->project);
 
         if (!$document) {
             return $this->reply(null, 404, "ERR-TASKS-NOT-FOUND");
         }
+
+        $this->can("update", $document);
+        $this->can("update", $task);
 
         $stackModel = new StackModel();
         $stack = $stackModel->find($task->stack);
@@ -167,139 +189,42 @@ class TasksController extends BaseController
             return $this->reply(null, 404, "ERR-TASKS-NOT-FOUND");
         }
 
-        $taskData = $this->request->getJSON();
-        helper("uuid");
-
-        $db = db_connect();
-        $db->transStart();
-
-        // generate list of new extensions
-        if (isset($taskData->extensions)) {
-            // Managing extensions
-            // delete the current task extensions
-            $taskExtensionModel = new TaskExtensionModel();
-            try {
-                if ($taskExtensionModel->where("task", $task->id)->delete() === false) {
-                    return $this->reply($taskExtensionModel->errors(), 500, "ERR-TASK-UPDATE");
-                }
-            } catch (\Exception $e) {
-                return $this->reply($e->getMessage(), 500, "ERR-TASK-UPDATE");
-            }
-
-            $extensions = array();
-            foreach ($taskData->extensions as $ext) {
-                $extension = new \stdClass();
-                $extension->task = $taskID;
-                $extension->title = $ext->title;
-                $extension->type = $ext->type;
-
-                if ($extension->type == "attachments") {
-                    $extension->content = "[]"; // just save a simple JSON array
-                } else {
-                    $extension->content = json_encode($ext->content);
-                }
-                $extension->options = json_encode($ext->options);
-
-                if (!isset($ext->id)) {
-                    $extension->id = uuid();  
-                } else {
-                    $extension->id = $ext->id;
-                }
-
-                $extensions[] = $extension;
-            }
-
-            if (count($extensions)) {    
-                try {
-                    if ($taskExtensionModel->insertBatch($extensions) === false) {
-                        return $this->reply($taskExtensionModel->errors(), 500, "ERR-TASK-UPDATE");    
-                    }
-                } catch (\Exception $e) {
-                    return $this->reply($e->getMessage(), 500, "ERR-TASK-UPDATE");
-                }
-            }
-        }
-
-        // generate a list of new assignees
-        if (isset($taskData->assignees)) {
-            // delete all assigned task users
-            $taskAssigneeModel = new TaskAssigneeModel();
-            try {
-                if ($taskAssigneeModel->where("task", $task->id)->delete() === false) {
-                    return $this->reply($taskAssigneeModel->errors(), 500, "ERR-TASK-UPDATE");
-                }
-            } catch (\Exception $e) {
-                return $this->reply($e->getMessage(), 500, "ERR-TASK-UPDATE");
-            }
-
-            $assignees = array();
-            foreach ($taskData->assignees as $person) {
-                $assignee = new \stdClass();
-                $assignee->task = $task->id;
-                $assignee->person = $person;
-                $assignees[] = $assignee;
-            }
-
-            // insert the assignees if any
-            if (count($assignees)) {
-                try {
-                    if ($taskAssigneeModel->insertBatch($assignees) === false) {
-                        return $this->reply($taskAssigneeModel->errors(), 500, "ERR-TASK-ASSIGNEES");    
-                    }
-                } catch (\Exception $e) {
-                    return $this->reply($e->getMessage(), 500, "ERR-TASK-ASSIGNEES");
-                }
-            }
-        }
-        
-        // convert tags to string
-        if (isset($taskData->tags)) {
-            $taskData->tags = json_encode($taskData->tags);
-        }
-
-        // convert repeats to string
-        if (isset($taskData->repeats)) {
-            $taskData->repeats = json_encode($taskData->repeats);
-        }
-
-        unset($taskData->id);
-        unset($taskData->position);
-        unset($taskData->project);
-        unset($taskData->stack);
-        unset($taskData->assignees);
-        unset($taskData->info);
-        $taskData->archived = null;
-
         // if somebody tries changing the owner and it's not the current owner then remove it
-        if ($taskData->owner && $taskData->owner != $user->id) {
+        if ($data->owner && $data->owner != $user->id) {
             return $this->reply(null, 403);
         }
 
         // if somebody tries changing the visibility (private, public) and it's not the owner then remove it
-        if ($taskData->public && $task->owner != $user->id) {
+        if ($data->public && $task->owner != $user->id) {
             return $this->reply(null, 403);
         }
 
-        // fix start date formatting
-        if (isset($taskData->startdate)) {
-            $taskData->startdate = substr(str_replace("T", " ", $taskData->startdate), 0, 19);
+        $this->db->transStart();
+
+        // Managing extensions
+        try {
+            $taskModel->addExtensions($task->id, $data->extensions);
+        } catch (\Exception $e) {
+            return $this->reply($e->getMessage(), 500, "ERR-TASK-UPDATE");
         }
 
-        // fix due date formatting
-        if (isset($taskData->duedate)) {
-            $taskData->duedate = substr(str_replace("T", " ", $taskData->duedate), 0, 19);
+        // generate a list of new assignees
+        try {
+            $taskModel->addAssignees($task->id, $data->extensions);
+        } catch (\Exception $e) {
+            return $this->reply($e->getMessage(), 500, "ERR-TASK-UPDATE");
         }
 
-        // fix completed date formatting
-        if (isset($taskData->completed)) {
-            $taskData->completed = substr(str_replace("T", " ", $taskData->completed), 0, 19);
-        }
-
-        if ($taskModel->update($taskID, $taskData) === false) {
+        // update the task
+        if ($taskModel->update($taskID, $data) === false) {
             return $this->reply(null, 500, "ERR-TASK-UPDATE");
         }
 
-        $db->transComplete();
+        $this->db->transComplete();
+
+        if ($this->db->transStatus() === false) {
+            return $this->reply(null, 500, "ERR-TASK-UPDATE");
+        }
 
         $this->addActivity(
             $document->id,
@@ -312,18 +237,25 @@ class TasksController extends BaseController
         return $this->reply(true);
     }
 
-    public function delete_v1($taskID)
+    public function delete_v1($taskId)
     {
-        $this->lock($taskID);
-
-        $taskModel = new TaskModel();
-        $task = $taskModel->find($taskID);
+        $this->lock($taskId);
+        $task = $this->getTask($taskId);
 
         if (!$task) {
             return $this->reply(null, 404, "ERR-TASKS-DELETE");
         }
 
+        $document = $this->getDocument($task->project);
+        if (!$document) {
+            return $this->reply(null, 404, "ERR-TASKS-DELETE");
+        }
+
+        $this->can("delete", $document);
+        $this->can("delete", $task);
+
         // delete selected task
+        $taskModel = new TaskModel();
         try {
             if ($taskModel->delete([$task->id]) === false) {
                 return $this->reply($taskModel->errors(), 500, "ERR-TASKS-DELETE");
@@ -343,29 +275,34 @@ class TasksController extends BaseController
         return $this->reply(true);
     }
 
-    public function get_watchers_v1($taskID)
+    public function get_watchers_v1($taskId)
     {
-        $user = $this->request->user;
-
-        helper("watchers");
-        $watchers = watchers_load($taskID, $user);
+        $userModel = new UserModel();
+        $userBuilder = $userModel->builder();
+        $usersQuery = $userBuilder->select("users.id, users.email, users.nickname, users.firstName, users.lastName, tasks_watchers.created")
+            ->join('tasks_watchers', 'tasks_watchers.user = users.id', 'left')
+            ->groupStart()
+                ->where('tasks_watchers.task', $taskId)
+                ->where('users.id !=', $this->request->user->id)
+            ->groupEnd()
+            ->orderBy('users.firstName', 'ASC')
+            ->get();
+        $watchers = $usersQuery->getResult();
 
         return $this->reply($watchers);
     }
 
-    public function add_watcher_v1($taskID)
+    public function add_watcher_v1($taskId)
     {
-        $user = $this->request->user;
-
         $taskWatcherModel = new TaskWatcherModel();
 
-        $watchers = $taskWatcherModel->where("task", $taskID)
-            ->where("user", $user->id)
+        $watchers = $taskWatcherModel->where("task", $taskId)
+            ->where("user", $this->request->user->id)
             ->findAll();
 
         $watcher = array(
-            "task" => $taskID,
-            "user" => $user->id
+            "task" => $taskId,
+            "user" => $this->request->user->id
         );
 
         if (!count($watchers)) {
@@ -390,7 +327,7 @@ class TasksController extends BaseController
         $this->addActivity(
             "",
             "", 
-            $taskID, 
+            $taskId, 
             $this::ACTION_CREATE, 
             $this::SECTION_WATCHER
         );
@@ -398,18 +335,16 @@ class TasksController extends BaseController
         return $this->reply(true);
     }
 
-    public function remove_watcher_v1($taskID)
+    public function remove_watcher_v1($taskId)
     {
-        $user = $this->request->user;
-
         $taskWatcherModel = new TaskWatcherModel();
 
         try {
             if ($taskWatcherModel
-                ->where("user", $user->id)
-                ->where("task", $taskID)->delete() === false
+                ->where("user", $this->request->user->id)
+                ->where("task", $taskId)->delete() === false
             ) {
-                return $this->reply($taskAssigneeModel->errors(), 500, "ERR-TASK-DELETE-WATCHER");
+                return $this->reply($taskWatcherModel->errors(), 500, "ERR-TASK-DELETE-WATCHER");
             }
         } catch (\Exception $e) {
             return $this->reply($e->getMessage(), 500, "ERR-TASK-DELETE-WATCHER");
@@ -418,7 +353,7 @@ class TasksController extends BaseController
         $this->addActivity(
             "",
             "", 
-            $taskID, 
+            $taskId, 
             $this::ACTION_DELETE, 
             $this::SECTION_WATCHER
         );
